+++ internal/usecase/errors.go
@@ -3,40 +3,12 @@ package usecase
 import "errors"
 
 var (
-	ErrUsernameTaken              = errors.New("this username is already taken") // 内部用エラー
-	ErrInvalidCredentials         = errors.New("invalid username or password")   // ログイン失敗
-	ErrUserIDMismatch             = errors.New("user ID mismatch")               // 権限エラー
-	ErrInvalidSession             = errors.New("session invalid or expired")     // セッション無効統合
-	ErrUserDeleted                = errors.New("user deleted")                   // ユーザー削除済み
-	ErrUserNotFound               = errors.New("user not found")                 // ユーザー未発見（Private含む）
-	ErrPlayerNotLinked            = errors.New("player not linked to user")      // プレイヤー未紐付
-	ErrUserPrivate                = errors.New("user profile is private")        // 内部用：非公開プロファイル
-	ErrIncorrectPassword          = errors.New("current password is incorrect")  // パスワード不一致
-	ErrInvalidPassword            = errors.New("invalid password")               // パスワード無効（詳細隠蔽）
-	ErrOperationFailed            = errors.New("operation failed")               // 操作失敗（詳細隠蔽）
-	ErrInternalError              = errors.New("internal error")                 // サーバ内部エラー（詳細隠蔽）
-	ErrInvalidPlayerName          = errors.New("invalid player name")            // プレイヤー名バリデーション失敗
-	ErrInvalidRecoveryCredentials = errors.New("invalid recovery credentials")   // リカバリー失敗（詳細隠蔽）
-	ErrUserAlreadyDeleted         = errors.New("user already deleted")           // 内部用
-	ErrUserNotDeleted             = errors.New("user not deleted")               // 内部用
+	ErrInvalidPlayerName = errors.New("invalid player name")
 
-	// ユーザー名バリデーションエラー
-	ErrUsernameEmpty       = errors.New("username cannot be empty")
-	ErrUsernameTooShort    = errors.New("username must be at least 5 characters")
-	ErrUsernameTooLong     = errors.New("username must be 50 characters or less")
-	ErrUsernameInvalidChar = errors.New("username can only contain lowercase letters and numbers")
+	ErrInvalidDifficulty = errors.New("invalid difficulty")
+	ErrChartNotFound     = errors.New("chart not found")
 
-	// パスワードバリデーションエラー
-	ErrPasswordTooShort = errors.New("password must be at least 8 characters")
-	ErrPasswordTooLong  = errors.New("password must be 128 characters or less")
+	ErrAdminRequired = errors.New("admin permission required")
 
-	// 難易度関連エラー
-	ErrInvalidDifficulty = errors.New("invalid difficulty") // 無効な難易度パラメータ
-	ErrChartNotFound     = errors.New("chart not found")    // 指定された難易度の譜面が存在しない
-
-	// 認可関連エラー
-	ErrAdminRequired = errors.New("admin permission required") // ADMIN権限が必要
-
-	// アプリバージョン関連エラー
-	ErrAppVersionUnsupported = errors.New("unsupported app version") // 対応していないアプリバージョン
+	ErrAppVersionUnsupported = errors.New("unsupported app version")
 )
