+++ internal/usecase/recovery_errors.go
@@ -1,0 +1,5 @@
+package usecase
+
+import "errors"
+
+var ErrInvalidRecoveryCredentials = errors.New("invalid recovery credentials")
